Adds an MVP for a team page with a map.
Pretty simple. Adds a map with the locations of our offices and locations
of remote workers with the same basic circle pictures below, grouped by
location. Click on a marker and a popup will show the name of the city and
you can click a link to see who works there. Very simple and "designed by
greg." Throwing it out there to start some discussion.